fix some stuff, slightly buggy

+++ include/Knob.h
@@ -7,20 +7,20 @@
 #include <cmath>
 
 class Knob : public SysState {
-public:
-    Knob(float upperBound, float lowerBound, float increment);
-    float getRotationISR();
-    float getRotation();
-    void updateRotation(std::string BA_curr);
+    public:
+        Knob(float upperBound, float lowerBound, float increment);
+        uint32_t getRotationISR();
+        float getRotation();
+        void updateRotation(std::string BA_curr);
 
-protected:
-    uint32_t rotationISR;
-    float rotation;
-    float rotationUpperBound;
-    float rotationLowerBound;
-    float increments;
-    std::string BA_prev;
-    bool incrementLast;
+    protected:
+        uint32_t rotationISR;
+        float rotation;
+        float rotationUpperBound;
+        float rotationLowerBound;
+        float increments;
+        std::string BA_prev;
+        bool incrementLast;
 };
 
 #endif
